requirements.txt
FlaskFlask
gunicorn
Flask-CORS